refactor: changed the directory structure
