+++ composeApp/src/commonMain/kotlin/io/github/drp08/studypal/presentation/App.kt
@@ -1,4 +1,4 @@
-package io.github.drp08.studypal
+package io.github.drp08.studypal.presentation
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.padding
@@ -8,9 +8,9 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import cafe.adriel.voyager.navigator.CurrentScreen
 import cafe.adriel.voyager.navigator.Navigator
-import io.github.drp08.studypal.navigation.BottomNavBar
-import io.github.drp08.studypal.navigation.BottomNavItem
-import io.github.drp08.studypal.screens.BlankScreen
+import io.github.drp08.studypal.presentation.navigation.BottomNavBar
+import io.github.drp08.studypal.presentation.navigation.BottomNavItem
+import io.github.drp08.studypal.presentation.screens.BlankScreen
 import org.jetbrains.compose.ui.tooling.preview.Preview
 
 @Composable
