+++ composeApp/src/commonMain/kotlin/io/github/drp08/studypal/presentation/navigation/BottomNavItem.kt
@@ -1,16 +1,14 @@
-package io.github.drp08.studypal.navigation
+package io.github.drp08.studypal.presentation.navigation
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.DateRange
 import androidx.compose.material.icons.filled.Home
 import androidx.compose.material.icons.filled.Person
-import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.vector.ImageVector
 import cafe.adriel.voyager.core.screen.Screen
-import io.github.drp08.studypal.screens.CalendarScreen
-import io.github.drp08.studypal.screens.HomeScreen
-import io.github.drp08.studypal.screens.PomodoroScreen
-import io.github.drp08.studypal.screens.ProfileScreen
+import io.github.drp08.studypal.presentation.screens.CalendarScreen
+import io.github.drp08.studypal.presentation.screens.HomeScreen
+import io.github.drp08.studypal.presentation.screens.ProfileScreen
 
 sealed class BottomNavItem(
     val label: String,
