+++ composeApp/src/commonMain/kotlin/io/github/drp08/studypal/presentation/screens/HomeScreen.kt
@@ -1,4 +1,4 @@
-package io.github.drp08.studypal.screens
+package io.github.drp08.studypal.presentation.screens
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -31,7 +31,7 @@ import cafe.adriel.voyager.navigator.LocalNavigator
 import cafe.adriel.voyager.navigator.Navigator
 import cafe.adriel.voyager.navigator.currentOrThrow
 import io.github.drp08.studypal.di.appModule
-import io.github.drp08.studypal.viewmodels.HomeViewModel
+import io.github.drp08.studypal.presentation.viewmodels.HomeViewModel
 import kotlinx.coroutines.delay
 import kotlinx.datetime.Clock
 import kotlinx.datetime.LocalTime
