+++ composeApp/src/commonMain/kotlin/io/github/drp08/studypal/presentation/screens/PomodoroScreen.kt
@@ -1,4 +1,4 @@
-package io.github.drp08.studypal.screens
+package io.github.drp08.studypal.presentation.screens
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
