+++ composeApp/src/iosMain/kotlin/MainViewController.kt
@@ -1,4 +1,4 @@
 import androidx.compose.ui.window.ComposeUIViewController
-import io.github.drp08.studypal.App
+import io.github.drp08.studypal.presentation.App
 
 fun MainViewController() = ComposeUIViewController { App() }
