--- server/src/main/kotlin/io/github/drp08/studypal/models/Schedule.kt
@@ -1,6 +1,0 @@
-package io.github.drp08.studypal.models
-
-import io.ktor.resources.Resource
-
-@Resource("/schedule")
-data class Schedule(val name: String = "Subject")
