+++ shared/src/commonMain/kotlin/io/github/drp08/studypal/utils/Constants.kt
@@ -2,5 +2,5 @@ package io.github.drp08.studypal.utils
 
 object Constants {
     const val SERVER_PORT = 8080
-    const val SERVER_HOST = "192.168.1.196"
+    const val SERVER_HOST = "146.169.162.11"
 }
